+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## master
 
+### Bugfix
+
+- [#7](https://github.com/trivago/babel-plugin-cloudinary/issues/7) Support asset names in `getBaseImageUrl` that are URL-encoded by Cloudinary API.
+
+### Features
+
 - Update dependencies based on npm's security audit
 
 ## [0.0.1](https://github.com/trivago/babel-plugin-cloudinary/tree/0.0.1)